fix: replace d2 with data engine
